requirements.txt
streamlit==1.29.0
sqlalchemy==2.0.23
pandas==2.1.3
python-dotenv==1.0.0

mysql-connector-python==8.1.0
psycopg2-binary==2.9.9
google-generativeai==0.3.1